raspberry/requirements.txt
python-socketio==5.7.2
RPi.GPIO==0.7.1
eventlet==0.33.3 